requirements.txt
flask
Pillow
imagehash
gunicorn
